requirements.txt
requests
schedule
google-generativeai
python-dotenv 